pyproject.toml
[project]
name = "latent-lab"
version = "0.1.0"
description = "ML/DL/LLM experimentation lab optimized for Apple Silicon"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Kiyeon Jeon" }]

dependencies = [
    # Core ML Frameworks
    "mlx>=0.22.0",
    "mlx-lm>=0.22.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "torchaudio>=2.5.0",

    # Classical ML
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",

    # Data
    "polars>=1.0.0",
    "duckdb>=1.0.0",
    "datasets>=3.0.0",
    "pyarrow>=17.0.0",

    # NLP & Transformers
    "transformers>=4.45.0",
    "tokenizers>=0.20.0",
    "accelerate>=1.0.0",
    "sentence-transformers>=3.0.0",

    # Experiment Management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.17.0",

    # Utilities
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
llm = [
    "ollama>=0.4.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
]
cv = [
    "ultralytics>=8.3.0",
    "opencv-python>=4.10.0",
    "coremltools>=8.0",
    "Pillow>=10.0.0",
]
dl = [
    "torchmetrics>=1.4.0",
    "lightning>=2.4.0",
    "timm>=1.0.0",
    "diffusers>=0.31.0",
    "safetensors>=0.4.0",
]
tuning = [
    "optuna>=4.0.0",
    "shap>=0.46.0",
]
rl = [
    "gymnasium[classic-control,box2d]>=1.0.0",
    "stable-baselines3[extra]>=2.4.0",
]
serving = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gradio>=5.0.0",
]
notebooks = [
    "marimo>=0.9.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "plotly>=5.24.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "pre-commit>=4.0.0",
    "ipykernel>=6.29.0",
]
all = [
    "latent-lab[llm,cv,dl,rl,tuning,serving,notebooks,dev]",
]

[project.scripts]
lab = "latent_lab.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/latent_lab"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "PTH"]

[tool.ruff.lint.isort]
known-first-party = ["latent_lab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
